Fix Carthage binary file reference naming
For binary frameworks distribution over Carthage, File referencing binaries should not be named Carthage.json.
To avoid naming confusion when using multiple framework binaries and allow the proper use of update command.

I didn't remove old Carthage.json in this commit and marked it as deprecated in the README file to let time to actual users to update their Cartfile.